+++ lib/rabarber/controllers/concerns/authorization.rb
@@ -12,18 +12,27 @@ module Rabarber
       def grant_access(action: nil, roles: nil, if: nil, unless: nil)
         dynamic_rule, negated_dynamic_rule = binding.local_variable_get(:if), binding.local_variable_get(:unless)
 
-        Permissions.write(self, action, roles, dynamic_rule, negated_dynamic_rule)
+        Rabarber::Permissions.add(
+          self,
+          Rabarber::Input::Actions.new(action).process,
+          Rabarber::Input::Roles.new(roles).process,
+          Rabarber::Input::DynamicRules.new(dynamic_rule).process,
+          Rabarber::Input::DynamicRules.new(negated_dynamic_rule).process
+        )
       end
     end
 
     private
 
     def verify_access
-      return if Permissions.access_granted?(
-        send(::Rabarber::Configuration.instance.current_user_method).roles, self.class, action_name.to_sym, self
+      Rabarber::Missing::Actions.new(self.class).handle
+      Rabarber::Missing::Roles.new(self.class).handle
+
+      return if Rabarber::Permissions.access_granted?(
+        send(Rabarber::Configuration.instance.current_user_method).roles, self.class, action_name.to_sym, self
       )
 
-      ::Rabarber::Configuration.instance.when_unauthorized.call(self)
+      Rabarber::Configuration.instance.when_unauthorized.call(self)
     end
   end
 end