+++ lib/rabarber/input/actions.rb
@@ -1,0 +1,30 @@
+# frozen_string_literal: true
+
+module Rabarber
+  module Input
+    class Actions < Rabarber::Input::Base
+      def initialize(
+        value,
+        error_type = Rabarber::InvalidArgumentError,
+        error_message = "Action name must be a Symbol or a String"
+      )
+        super
+      end
+
+      private
+
+      def valid?
+        (value.is_a?(String) || value.is_a?(Symbol)) && value.present? || value.nil?
+      end
+
+      def processed_value
+        case value
+        when String, Symbol
+          value.to_sym
+        when nil
+          value
+        end
+      end
+    end
+  end
+end