+++ lib/rabarber/models/concerns/has_roles.rb
@@ -5,7 +5,9 @@ module Rabarber
     extend ActiveSupport::Concern
 
     included do
-      raise Error, "Rabarber::HasRoles can only be included once" if defined?(@@included) && @@included != name
+      if defined?(@@included) && @@included != name
+        raise Rabarber::Error, "Rabarber::HasRoles can only be included once"
+      end
 
       @@included = name
 
@@ -19,26 +21,30 @@ module Rabarber
     end
 
     def has_role?(*role_names)
-      (roles & RoleNames.pre_process(role_names)).any?
+      (roles & process_role_names(role_names)).any?
     end
 
     def assign_roles(*role_names, create_new: true)
-      roles_to_assign = RoleNames.pre_process(role_names)
+      roles_to_assign = process_role_names(role_names)
 
       create_new_roles(roles_to_assign) if create_new
 
-      rabarber_roles << Role.where(name: roles_to_assign) - rabarber_roles
+      rabarber_roles << Rabarber::Role.where(name: roles_to_assign) - rabarber_roles
     end
 
     def revoke_roles(*role_names)
-      self.rabarber_roles = rabarber_roles - Role.where(name: RoleNames.pre_process(role_names))
+      self.rabarber_roles = rabarber_roles - Rabarber::Role.where(name: process_role_names(role_names))
     end
 
     private
 
     def create_new_roles(role_names)
-      new_roles = role_names - Role.names
-      new_roles.each { |role_name| Role.create!(name: role_name) }
+      new_roles = role_names - Rabarber::Role.names
+      new_roles.each { |role_name| Rabarber::Role.create!(name: role_name) }
+    end
+
+    def process_role_names(role_names)
+      Rabarber::Input::Roles.new(role_names).process
     end
   end
 end